modified:   __pycache__/__init__.cpython-311.pyc 	modified:   __pycache__/layer.cpython-311.pyc 	modified:   __pycache__/metrics.cpython-311.pyc 	modified:   __pycache__/nn.cpython-311.pyc 	new file:   __pycache__/optimizer.cpython-311.pyc 	new file:   example.py 	modified:   nn.py 	new file:   optimizer.py
